Remove redundant code, reuse code
