+++ list_node.rb
@@ -1,0 +1,8 @@
+class ListNode
+  attr_accessor :value, :next_node
+
+  def initialize(value, next_node = nil)
+    @value = value
+    @next_node = next_node
+  end
+end
