+++ queue.rb
@@ -1,159 +1,39 @@
-class Node
-  attr_accessor :value, :next_node
+require_relative "./linked_list"
 
-  def initialize(value, next_node = nil)
-    @value = value
-    @next_node = next_node
-  end
-end
-
-module LinkedList
-  # setup head and tail
-  @head = nil
-  @tail = nil
-
-  def add(number)
-    new_node = Node.new(number)
-    if @head.nil?
-      @head = new_node
-    else
-      # insert the node at the end of the list / tail
-      @tail.next_node = new_node
-    end
-    @tail = new_node
-  end
-
-  def get(index)
-    next_node = @head
-    idx = 0
-    while idx < index
-      next_node = next_node.next_node
-      idx += 1
-    end
-
-    next_node.value
-  end
-
-  def print_all
-    next_node = @head
-    idx = 0
-    until next_node.nil?
-      p idx.to_s + ': ' + next_node.value.to_s
-      next_node = next_node.next_node
-      idx += 1
-    end
-  end
-
-  def add_at(index, item)
-    new_node = Node.new(item)
-
-    # Either we are adding at the front, in the middle,
-    # or at the back of the list
-
-    if index.zero?
-      # Add the new node to the front of the list
-      new_node.next_node = @head
-      @head = new_node
-    else
-      next_node = @head
-      idx = 0
-
-      # Traverse until the node just before the insert index
-      while idx < index - 1
-        next_node = next_node.next_node
-        idx += 1
-      end
-
-      # idx = index - 1
-      # insert the new node after this position
-
-      # Check if the next node is nil, meaning we're inserting at
-      #  the end of the list
-      if next_node.next_node.nil?
-        # insert the node at the end of the list / tail
-        next_node.next_node = new_node
-        @tail = new_node
-      else
-        new_node.next_node = next_node.next_node
-        next_node.next_node = new_node
-      end
-    end
-  end
-
-  def remove_at(index)
-    # Either we are removing from the front, in the middle,
-    #  or at the back of the list
-
-    if index.zero?
-      # Remove the node from the front of the list
-      new_head = @head.next_node
-      @head.next_node = nil
-      @head = new_head
-    else
-      next_node = @head
-      idx = 0
-
-      # Traverse until the node just before the delete index
-      while idx < index - 1
-        next_node = next_node.next_node
-        idx += 1
-      end
-
-      # idx = index - 1
-      # delete the node after this position
-
-      # Check if the next node is nil, meaning we're deleting a
-      #  node which does not exist
-      if next_node.next_node.nil?
-        puts "Can't delete an inexistent node"
-
-      # Check if the next node is the last node,
-      #  meaning we're deleting at the end of the list
-      elsif next_node.next_node.next_node.nil?
-        next_node.next_node = nil
-        @tail = next_node
-      else
-        next_node.next_node = next_node.next_node.next_node
-
-      end
-    end
-  end
-end
-
-class Queue
-  include LinkedList
+class MyQueue < LinkedList
+  # include LinkedList
 
   # We add items to the queue at the back of the list
   # The add method already exists in the linked list
 
-  def remove
+  def take_off
     # We pop items off the queue from the front of the list(index 0)
     item = -1
     unless @head.nil?
       item = get(0)
-      remove_at(0)
+      remove(0)
     end
     item
   end
 end
 
-queue = Queue.new
+queue = MyQueue.new
 
 queue.add(3)
 queue.add(5)
-puts queue.remove
+puts queue.take_off
 # => 3
 
 queue.add(2)
 queue.add(7)
-puts queue.remove
+puts queue.take_off
 # => 5
 
-puts queue.remove
+puts queue.take_off
 # => 2
 
-puts queue.remove
+puts queue.take_off
 # => 7
 
-puts queue.remove
+puts queue.take_off
 # => -1
